requirements.txt
PySide6>=6.6.0,<7
Pillow>=9.0.0,<10